+++ bindings/go/plugin/go.mod
@@ -1,0 +1,29 @@
+module ocm.software/open-component-model/bindings/go/plugin
+
+go 1.24.1
+
+require (
+	github.com/invopop/jsonschema v0.13.0
+	github.com/santhosh-tekuri/jsonschema/v6 v6.0.1
+	github.com/stretchr/testify v1.10.0
+	ocm.software/open-component-model/bindings/go/descriptor/runtime v0.0.0-20250410115029-6dcf800db82b
+	ocm.software/open-component-model/bindings/go/oci v0.0.0-20250414115417-bc063a2d4cca
+	ocm.software/open-component-model/bindings/go/runtime v0.0.0-20250410115029-6dcf800db82b
+)
+
+require (
+	github.com/bahlo/generic-list-go v0.2.0 // indirect
+	github.com/buger/jsonparser v1.1.1 // indirect
+	github.com/cyberphone/json-canonicalization v0.0.0-20241213102144-19d51d7fe467 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
+	github.com/opencontainers/go-digest v1.0.0 // indirect
+	github.com/opencontainers/image-spec v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/wk8/go-ordered-map/v2 v2.1.8 // indirect
+	golang.org/x/text v0.21.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+	ocm.software/open-component-model/bindings/go/descriptor/v2 v2.0.0-20250410115029-6dcf800db82b // indirect
+	oras.land/oras-go/v2 v2.5.0 // indirect
+	sigs.k8s.io/yaml v1.4.0 // indirect
+)
