+++ bindings/go/plugin/manager/component_version_repository_registry_test.go
@@ -1,0 +1,41 @@
+package manager
+
+import (
+	"context"
+	"log/slog"
+	"os"
+	"runtime"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+	v2 "ocm.software/open-component-model/bindings/go/oci/spec/repository/v1"
+)
+
+func TestGetTransferPlugin(t *testing.T) {
+	r := require.New(t)
+	testctx := context.Background()
+	pm := NewPluginManager(testctx, slog.New(slog.DiscardHandler))
+	location := "testdata/darwin"
+	if runtime.GOOS == "linux" {
+		location = "testdata/linux"
+	}
+	err := pm.RegisterPluginsAtLocation(testctx, location, WithIdleTimeout(10*time.Second))
+	r.NoError(err)
+	tmp, err := os.CreateTemp("", "test.file")
+	r.NoError(err)
+	t.Cleanup(func() {
+		r.NoError(os.RemoveAll("/tmp/ocm_plugin_generic.sock"))
+		r.NoError(pm.Shutdown(testctx))
+		r.NoError(tmp.Close())
+		r.NoError(os.Remove(tmp.Name()))
+	})
+
+	got, err := GetReadWriteComponentVersionRepository(testctx, &v2.OCIRepository{}, WithPluginManager(pm))
+	r.NoError(err)
+	r.NoError(got.GetLocalResource(testctx, GetLocalResourceRequest[*v2.OCIRepository]{
+		TargetLocation: Location{
+			Value: tmp.Name(),
+		},
+	}, nil))
+}
