+++ bindings/go/plugin/manager/generic/README.md
@@ -1,0 +1,1 @@
+# Generic Plugin for testing
