+++ bindings/go/plugin/manager/manager_test.go
@@ -1,0 +1,7 @@
+package manager
+
+import "testing"
+
+func TestPluginManager(t *testing.T) {
+	
+}
