--- plugins/manager/doc.go
@@ -1,2 +1,0 @@
-// Package manager provides functionality for
-package manager
