--- plugins/manager/generic/README.md
@@ -1,1 +1,0 @@
-# Generic Plugin for testing
