--- plugins/manager/internal/README.md
@@ -1,3 +1,0 @@
-# Internal Plugin
-
-This plugin tests the functionality of being able to register an internal implementation for a plugin.
