--- plugins/manager/manager_test.go
@@ -1,7 +1,0 @@
-package manager
-
-import "testing"
-
-func TestPluginManager(t *testing.T) {
-	
-}
